Reorganization into src folder
