+++ jsdoc.json
@@ -1,6 +1,6 @@
 {
     "source": {
-        "include": ["."],
+        "include": ["./src"],
         "includePattern": ".js$",
         "excludePattern": "(node_modules/|documentation|tests)"
     },
