+++ src/client/aws/baseCredentialProvider.js
