+++ src/client/aws/index.js
