+++ src/client/aws/simpleCredentialProvider.js
