+++ src/client/aws/ssmCredentialProvider.js
