+++ src/client/simple.js
